move generic _run() to wamp.component
